Change some form handling rules
jsx-a11y/label-has-for is deprecated in future versions of eslint so we should disable it now in favor of label-as-associated control
https://github.com/evcohen/eslint-plugin-jsx-a11y/blob/master/docs/rules/label-has-for.md

fs-styles currently does not support nesting, so the default `assert` setting of both is too strict. Switching to `either` allows either nesting or using a `for` or `htmlFor` parameter. Until we move off of fs-styles this will prevent many spurious linting issues.
https://github.com/evcohen/eslint-plugin-jsx-a11y/blob/master/docs/rules/label-has-associated-control.md